NXT-129 Move project updates to own component. Add null for reasonDecisionText prop
